% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{addCloneToCompleted}
\alias{addCloneToCompleted}
\title{Function to add a clone ID and its size to a file.}
\usage{
addCloneToCompleted(folder, cloneID, size)
}
\arguments{
\item{folder}{The path to the folder, where the file lies (filename is always "clones_completed.csv").}

\item{cloneID}{The clone ID to be added.}

\item{size}{The size of the clone.}
}
\description{
Function to add a clone ID and its size to a file.
}
