% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignmentValidCodons.R
\name{alignmentValidCodons}
\alias{alignmentValidCodons}
\title{Function to ensure, that a given sequence alignment has valid codons, i.e. a number of bases divisible by 3 without rest.}
\usage{
alignmentValidCodons(
  inputAlignment,
  cutEnd = TRUE,
  removeSTOPcodons = TRUE,
  maxSTOPperSequence = 1,
  do.warn = FALSE
)
}
\arguments{
\item{inputAlignment}{Input alignment (FASTA format in \link[bio3d]{bio3d}) with nucleic acid sequences.}

\item{cutEnd}{If \code{TRUE}, surplus bases will be cut at the end otherwise at the beginning of the nucleotide sequences (default: \code{TRUE}).}

\item{removeSTOPcodons}{If \code{TRUE}, the first position in STOP codons in the sequences is replaced by a dash (default: \code{TRUE}).}

\item{maxSTOPperSequence}{If "removeSTOPcodons" is \code{TRUE} and the number of STOP codons for a sequence exceeds this very value then this particular sequence is removed from the alignment (default: 1).}

\item{do.warn}{If \code{TRUE}, a warning will issued if the aligned sequences do not have a length which is a clean multiple of 3 (default: \code{FALSE}).}
}
\value{
A list with the two elements "alignment" (the modified alignment) and "report", the latter being a named list with the elements "noCutPositions" (number of positions, that have been cut) and the list "STOPreplaced". List "STOPreplaced" holds identifiers for the sequences and "noSTOPreplaced", the number of STOP codons which have been replaced (only for those where "noSTOPreplaced" is greater than zero).
}
\description{
Function to ensure, that a given sequence alignment has valid codons, i.e. a number of bases divisible by 3 without rest.
}
\examples{
\dontrun{alignmentValidCodons <- function( inputAlignment = ali )}

}
