% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloneArborescence.R
\name{cloneArborescence}
\alias{cloneArborescence}
\title{Function to calculate the arborescence for a fully connected graph of B cell clone members.}
\usage{
cloneArborescence(
  clone,
  cloneID,
  columnSeqID,
  columnSubclass,
  tree,
  outputFolder,
  plotFormat = "png",
  colourSpecification = list(colours = list(M = "#FF0000", IgA1 = "#9999CF", IgA2 =
    "#000088", IgG1 = "#326600", IgG2 = "#58B200", IgG3 = "#8BFF19", IgG4 = "#CBFF99"),
    germlineColour = "#AAAAAA")
)
}
\arguments{
\item{clone}{Matrix / data frame holding the clone information.}

\item{cloneID}{Identifier of clone.}

\item{columnSeqID}{Name of the column holding the ID of the observation.}

\item{columnSubclass}{Name of the column holding the isotype subclass (see manual page of function \link{getArborescence} for a list of allowed levels in this column).}

\item{tree}{Character, filename of the phylogenetic tree.}

\item{outputFolder}{Folder, where the arborescence data and plot are to be stored.}

\item{plotFormat}{Either 'png' or 'pdf'. (default: 'png')}

\item{colourSpecification}{List of colours, a meaningful default is set.}
}
\value{
A list with the following elements:
\describe{
  \item{graph}{\code{igraph} object holding the arborescence tree.}
  \item{table}{A data.frame object holding the edgelist of the arborescence tree.}
  \item{lineage_plot}{\code{ggplot} object of a lineage plot where sequences are ordered by their distance from the germline and coloured by isotypes. See vignette for more details.}
  \item{csr_events}{data.frame showing edges in the tree which switches isotype, and an estimated distance-from-germline of this switching event. This is defaulted as the mean distance-from-germline of the two nodes connected by the said edge.}
}
}
\description{
Function to calculate the arborescence for a fully connected graph of B cell clone members.
}
