% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{convertDNAStringSetToList}
\alias{convertDNAStringSetToList}
\title{Function to parse DNAStringSet into nested list}
\usage{
convertDNAStringSetToList(DNAStrings)
}
\arguments{
\item{DNAStrings}{object of class \code{DNAStringsSet}, holding the sequences. Can be obtained by \code{Biostrings::readDNAstringSet( fasta )}, where 'fasta' is the FASTA file holding the sequences.}
}
\value{
A list of list, the first layer organised by species and the second layer organised by allele.
}
\description{
This function parses user-supplied sequences to give a 'nested list' of sequences as desired by \code{cloneLineage} to
fetch germline sequence to perform lineage inference. NOTE: it is assumed sequences are directly fetched from IMGT and obeys the FASTA header format imposed by IMGT!
}
