% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{curDateTime}
\alias{curDateTime}
\title{Function to print the current date and the time.}
\usage{
curDateTime()
}
\value{
A string with the current datetime in a specific format.
}
\description{
Function to print the current date and the time.
}
