% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doBatchCloneAnalysis.R
\name{doBatchCloneAnalysis}
\alias{doBatchCloneAnalysis}
\title{Perform phylogenetic inference & tree analysis in a batch}
\usage{
doBatchCloneAnalysis(
  inputDF,
  outputFolder,
  species = "Human",
  minCloneSize = 3,
  sequence_column = "Sequence",
  cloneID_column = "CloneID",
  label_column = "Seq_ID",
  colourColumn = "Subclass",
  IGHVgeneandallele_column = "V.GENE.and.allele",
  germlineSet = NULL,
  plotFormat = "png",
  phyloTreeType = "simple",
  phyloTreeOptions = NULL,
  makeArboTree = TRUE,
  useTempDir = TRUE
)
}
\arguments{
\item{inputDF}{input data.frame containing repertoire data.}

\item{outputFolder}{character, path for the folder where output files are to be stored.}

\item{species}{characer, species. For now only "Human" is supported and accepted.}

\item{minCloneSize}{numeric, minimum clone size to be considered for tree inference. (default: 3)}

\item{sequence_column}{character, column name in \code{inputDF} which holds the sequences. (default: "Sequence")}

\item{cloneID_column}{character, column name in \code{inputDF} which holds the clone IDs. (default: "CloneID")}

\item{label_column}{character, column name in \code{inputDF} which holds the sequence identifiers. (default: "Seq_ID")}

\item{colourColumn}{character, column name in \code{inputDF} which is to be categorised by colours in plots. (default: "Subclass")}

\item{IGHVgeneandallele_column}{character, column name in \code{inputDF} which holds annotated germline V gene names. (default: "V.GENE.and.allele")}

\item{germlineSet}{Optional, filepath pointing to a FASTA file containing germline sequences to be used. (default: NULL)}

\item{plotFormat}{Either 'png' or 'pdf' (default: 'png')}

\item{phyloTreeType}{"simple", "dnapars" or "igphyml". "simple" refers to a neighbour-joining tree; "dnapars" constructs a maximum parsimony tree using the phylip "dnapars" program (a local installation of the program is required). "igphyml" constructs a maximum likelihood tree taking into account mutational hotspot contexts in immunoglobulins, but is the most time consuming. (default: "dnapars")}

\item{phyloTreeOptions}{A list to be fed as the 'parameter' entry in the \code{treeConstruction} argument of the \code{cloneLineage} function (see documentation of the \code{cloneLineage} function). If \code{NULL}, default settings will be used. (default: \code{NULL})}

\item{makeArboTree}{Should arborescence tree be calculated? (default: TRUE)}

\item{useTempDir}{If \code{TRUE}, generate temporary directory and write results there. (default: \code{TRUE})}
}
\value{
A list with each element named by a clone ID, each itself a list with two elements:
\describe{
  \item{distances}{data.frame with the distance-from-germline calculated from the lineage trees for each sequence in the clone. See Vignette for details.}
  \item{csr_events}{data.frame listing all class-switching events and an estimate distance-from-germline at which such event takes place.}
}
}
\description{
This function loops through clones and perform IgPhyML phylogenetic tree inference and arborescence tree construction, and calculates tree metrics, as detailed in vignette of this package.
}
