% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/germlineDistances.R
\name{GermlineLikeness}
\alias{GermlineLikeness}
\title{Function to calculate germline likeness of a pre-selected subset of clones}
\usage{
GermlineLikeness(tb)
}
\arguments{
\item{tb}{data.frame, subset of the output from \code{summariseGermlineDistance} corresponding to a set of clones of interest.}
}
\value{
numeric, Germline Likeness metric (between 0 and 1) summarising the given set of clones.
}
\description{
Internal function used by \code{getGermlineLikeness}.
}
