% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getArborescence.R
\name{getArborescence}
\alias{getArborescence}
\title{Function to transform a distance matrix of observations into a list of "Arc"s, a tuple defined in the python script "edmond_arborescence.py"}
\usage{
getArborescence(distances, dfSubclass, germline)
}
\arguments{
\item{distances}{Either a "dist" or a matrix with the observations.}

\item{dfSubclass}{A two-column data frame holding:
\describe{
  \item{SeqID}{The sequence ID as used in the distance object.}
  \item{Class}{The isotype subclass of the observations.}
}}

\item{germline}{character, the sequence to be taken as the germline
Note, that the germline has subclass type "germline" and observations without a valid subclass must be labelled 'NA' (will be ignored).}
}
\value{
A data frame with the columns "startID" (number of start), "endID" (number of end point), "weight" (weight of the arc), "startLabel" (the labels for the start IDs) and "endID" (the labels for the end IDs) for an arborescence minimum spanning tree.
}
\description{
Function to transform a distance matrix of observations into a list of "Arc"s, a tuple defined in the python script "edmond_arborescence.py"
}
