% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCloneIDsOrdered.R
\name{getCloneIDsOrdered}
\alias{getCloneIDsOrdered}
\title{Function to analyse B cell clone lineages, by producing an alignment and infering a phylogenetic tree.}
\usage{
getCloneIDsOrdered(input, cloneIdColumn, decreasing = TRUE)
}
\arguments{
\item{input}{Matrix / data frame with values.}

\item{cloneIdColumn}{Name of the column holding the clone IDs.}

\item{decreasing}{If \code{TRUE}, the clones will be ordered according to ever decreasing sizes.}
}
\description{
Function to analyse B cell clone lineages, by producing an alignment and infering a phylogenetic tree.
}
\examples{
\dontrun{cloneIDdf <- getCloneIDsOrdered( input = inputDF, cloneIdColumn = "Unique_CloneID" )}

}
