% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{getColourVectorFromLevels}
\alias{getColourVectorFromLevels}
\title{Function to generate a vector of colours from a vector of levels and a colour specification list}
\usage{
getColourVectorFromLevels(levels, colourSpecification)
}
\arguments{
\item{levels}{Vector of levels.}

\item{colourSpecification}{List that specifies which levels are to be represented by which respective colour. Make sure, that the definition matches the levels in parameter \code{levels}.}
}
\value{
A vector of colours in the order of the input levels.
}
\description{
Function to generate a vector of colours from a vector of levels and a colour specification list
}
