% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{getCSEvents}
\alias{getCSEvents}
\title{Function to parse class-switching events from arborescence tree analysis}
\usage{
getCSEvents(
  clone,
  columnSeqID,
  columnSubclass,
  dfArborescence,
  germline,
  path_stats
)
}
\arguments{
\item{clone}{data.frame of clone information.}

\item{columnSeqID}{column name holding Sequence IDs.}

\item{columnSubclass}{column name holding subclasses information.}

\item{dfArborescence}{data.frame of arborescence tree edge list.}

\item{germline}{Root ID for germline.}

\item{path_stats}{data.frame of distances from germline calculated for the arborescence tree.}
}
\value{
A data.frame.
}
\description{
This function parses edges in the arborescence tree that switches isotype/subisotype, 
and offer a distance-from-germline estimate of this switching event, as the mean distance-from-germline 
of the two ends of each of such edge.
}
