% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDistancesFromTree.R
\name{getDistancesFromTree}
\alias{getDistancesFromTree}
\title{Function to calculate the distances for a B cell clone lineage from a tree object.}
\usage{
getDistancesFromTree(tree, digits = 4)
}
\arguments{
\item{tree}{Input tree, type \code{phylo}.}

\item{digits}{Number of digits to consider (default: 4).}
}
\description{
Function to calculate the distances for a B cell clone lineage from a tree object.
}
