% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/germlineDistances.R
\name{getDistPercentile}
\alias{getDistPercentile}
\title{Function to calculate percentile of distance-from-germline over all clones in the selection}
\usage{
getDistPercentile(dists, percentiles)
}
\arguments{
\item{dists}{numeric vector, median distance-from-germline of a set of clones.}

\item{percentiles}{numeric vector, a list of quantiles to consider. For percentiles, they should be defined between 0 and 1 with step 0.01.}
}
\value{
numeric vector, percentile in the distance-from-germline distribution over all clones, for each corresponding entry in \code{dist}.
}
\description{
Internal function used by \code{summariseGermlineDistance}.
}
