% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/germlineDistances.R
\name{getGermlineDistance}
\alias{getGermlineDistance}
\title{Function to calculate distance from germline for all observations in a given lineage tree}
\usage{
getGermlineDistance(tree_file)
}
\arguments{
\item{tree_file}{Character, filename of the phylogenetic tree in newick format.}
}
\value{
A data.frame with two columns with the following names:
\describe{
  \item{SeqID}{Character, Identifier of sequence given in the tree file.}
  \item{distFromGermline}{Numeric, distance from the identified germline (i.e. total lengths of edges traversed in order to get from the germline to the given sequence.}
}
}
\description{
Read a newick tree to calculate distance-from-germline. Assume germline is in the tree, with name beginning with 'IG.V' where '.' can be any character.
}
