% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/germlineDistances.R
\name{getGermlineLikeness}
\alias{getGermlineLikeness}
\title{Function to summarise Germline Likeness for repertoires}
\usage{
getGermlineLikeness(
  distFromGermline_summary,
  metadata_table,
  summarise_variables
)
}
\arguments{
\item{distFromGermline_summary}{data.frame, output of \code{summariseGermlineDistance}. With metadata annotation including columns given in \code{summarise_variables}.}

\item{metadata_table}{data.frame with meta data, 1 row per sample/donor as appropriate (see Example).}

\item{summarise_variables}{Character vector, names of columns in \code{distFromGermline} used to partition the sequences into subsets for which summary is sought}
}
\value{
A data.frame identical to \code{metadata_table} with the additional column "GermlineLikeness" holding the Germline Likeness metric (numeric) of lineages matching the given combination of metadata attributes
}
\description{
The function loops through each row in \code{metadata_table} and select the subset of lineages from \code{distFromGermline_summary} relevant to the given metadata combination,
and calculate the Germline Likeness metric for this given set of lineages.
}
\examples{
\dontrun{
# We have included in the package the pre-computed germline distances of 
# sequences in the 'input' data frame, here use this as an example
distFromGermline <- system.file( "extdata/input_GermlineDistances.csv", package = "BrepPhylo")
distFromGermline <- read.csv(distFromGermline)

# set up a table containing metadata with which you want to
# obtain a summary of Germline Likeness
# here simply set up one column with PatientID
patient_details <- data.frame( PatientID = unique( distFromGermline[, c("PatientID") ] ) )

germline_likeness <- getGermlineLikeness( 
  germlineDistance_summary,
  metadata_table = patient_details,
  summarise_variables = "PatientID" 
)
germline_likeness
}

}
