% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IgPhyML.R
\name{IgPhyML}
\alias{IgPhyML}
\title{Function that calls IgPhyML to calculate a phylogenetic tree and returns it including meta-data.}
\usage{
IgPhyML(
  path,
  root,
  accuracy = "basic",
  executable = "",
  log = NULL,
  useTempDir = TRUE
)
}
\arguments{
\item{path}{Input FASTA file path.}

\item{root}{Root ID (germline sequence).}

\item{accuracy}{Either "basic", "high" or "extreme", this parameter affects the internal calls of IgPhyML which will affect runtime considerably (default: "basic").}

\item{executable}{If not empty, this file will be used for the call of \code{IgPhyML} otherwise the function will attempt to find the executable (default: "").}

\item{log}{If not \code{NULL}, log entries over the execution will be written to the given connection.}

\item{useTempDir}{If \code{TRUE}, generate temporary directory and write results there. (default: \code{TRUE})}
}
\value{
A list holding the intermediary and final trees and additional data (elements: "success", "topology", "fit" )/
}
\description{
Note, that what this function implements is a two-step optimization: First, the topology is searched for using M0/GY49 (\href{https://doi.org/10.1093/oxfordjournals.molbev.a040153}{Goldman and Yang (1994)}; \href{https://doi.org/10.1016/S0169-5347(00)01994-7}{Yang and Bielawski (2000)}), then this topology is fixed and used 
as starting point for the fitting by the HLP17 model (specialized for antibody lineages \href{https://doi.org/10.1534/genetics.116.196303}{Hoehn, Lunter and Pybus (2017)}). One could further optimize the topology by specifying "-o tlr" instead of "-o lr".
}
\examples{
\dontrun{IgPhyML( path = "/path/to/input.fasta", root = "V4-59" )}

}
