% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.valid.IsotypeSwitch.R
\name{is.valid.IsotypeSwitch}
\alias{is.valid.IsotypeSwitch}
\title{Function that checks, whether a phylogenetic relationship in B cells is possible or not according to the genetic arrangement.}
\usage{
is.valid.IsotypeSwitch(subclassParent, subclassChild)
}
\arguments{
\item{subclassParent}{The isotype subclass of the possible parent.}

\item{subclassChild}{The isotype subclass of the possible child.}
}
\value{
Boolean, indicating whether this parent - child relationship is possible (\code{TRUE}) or not (\code{FALSE}). If both subclasses are the same, the result is \code{TRUE}. If the input subclasses are not valid, 'NA' is returned.
}
\description{
The genetic order in humans is "germline", 'M', 'D', "G3", "G1", "A1", "G2", "G4", 'E' and "A2". Note, that the letters "Ig" are cut off the input, in case they are present.
}
\examples{
# FALSE
is.valid.IsotypeSwitch( "IgG3", "IgM" )

# TRUE
is.valid.IsotypeSwitch( "A1", "A2" )

}
