% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeAlignmentPDF.R
\name{makeAlignmentPDF}
\alias{makeAlignmentPDF}
\title{Function to generate PDF files holding an alignment using the function \link{msaPrettyPrint_extended} and \code{LaTeX}.}
\usage{
makeAlignmentPDF(clone, alignment, cloneIdColumn, outputFolder, colours)
}
\arguments{
\item{clone}{Clone dataframe.}

\item{alignment}{Alignment for the clone (including the germline sequence at the first position).}

\item{cloneIdColumn}{Name of the column holding the clone IDs.}

\item{outputFolder}{Path to the output folder.}

\item{colours}{A named vector of HTML colours for the sequences (including the '#' sign).}
}
\description{
Function to generate PDF files holding an alignment using the function \link{msaPrettyPrint_extended} and \code{LaTeX}.
}
\examples{
\dontrun{makeAlignmentPDF <- function( clone = clone,
                              alignment = alignment,
                              cloneIdColumn = "Unique_CloneID",
                              outputFolder = "~/Desktop",
                              colours = colours )}

}
