% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msaPrettyPrint_extended.R
\name{msaPrettyPrint_extended}
\alias{msaPrettyPrint_extended}
\title{Function extending the original "msa::msaPrettyPrint" function.}
\usage{
msaPrettyPrint_extended(
  x,
  y,
  output = c("pdf", "tex", "dvi", "asis"),
  subset = NULL,
  file = NULL,
  alFile = NULL,
  askForOverwrite = TRUE,
  psFonts = FALSE,
  vecAdditionalHeader = c(),
  code = NA,
  paperWidth = 11,
  paperHeight = 8.5,
  margins = c(0.1, 0.3),
  shadingMode = c("identical", "similar", "functional"),
  shadingModeArg = NA,
  shadingColors = c("blues", "reds", "greens", "grays", "black"),
  showConsensus = c("bottom", "top", "none"),
  consensusColors = c("ColdHot", "HotCold", "BlueRed", "RedBlue", "GreenRed",
    "RedGreen", "Gray"),
  consensusThreshold = 50,
  showLogo = c("top", "bottom", "none"),
  logoColors = c("chemical", "rasmol", "hydropathy", "structure", "standard area",
    "accessible area"),
  showLogoScale = c("none", "leftright", "left", "right"),
  showNames = c("left", "right", "none"),
  showNumbering = c("right", "left", "none"),
  showLegend = TRUE,
  furtherCode = NA,
  verbose = FALSE
)
}
\arguments{
\item{...}{As the original one.

This function should not be used directly.}
}
\description{
Function extending the original "msa::msaPrettyPrint" function.
}
