% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSR.R
\name{summariseCSR}
\alias{summariseCSR}
\title{Function to summarise details of different class-switch recombination events}
\usage{
summariseCSR(
  tree_stats,
  cloneID_column = "CloneID",
  dist_column = "dist",
  summarise_variables
)
}
\arguments{
\item{tree_stats}{data.frame, output of \code{getSummaryFromBatch} (specifically the element \code{csr_events} from the output list), and annotated with extra columns detailing metadata of samples.}

\item{cloneID_column}{Character, column name in \code{getSummaryFromBatch} which holds the clone ID to which the given sequence belongs. (Default: 'CloneID')}

\item{dist_column}{Character, column name in \code{getSummaryFromBatch} which holds the distance-from-germline metrics. (default: 'dist')}

\item{summarise_variables}{Character vector, names of columns in \code{getSummaryFromBatch} used to partition the sequences into subsets for which summary is sought.}
}
\value{
A data.frame with each row corresponding to one specific CSR switch (i.e. distinct combination of start- and end-points of the switch) in different subset of the data defined by the \code{summarise_variables} column(s). The following summaries were sought and stored in separate columns:
\describe{
  \item{nEvent}{Numeric, total number of switches involving the two named isotypes in the dataset subset in question.}
  \item{meanDist_from_germline}{Numeric, mean distance-from-germline of CSR events involving the two named isotypes.}
  \item{nClone}{Numeric, total number of clonotypes in the dataset subset (should be the same across all rows for the same subset of data).}
  \item{nClone_events}{Numeric, between 0 and 1, the proportion of clones with the named CSR switch observed.}
}
}
\description{
The function summarise the CSR details from \code{getSummaryFromBatch}.
}
\examples{
\dontrun{
# we provide doBatchCloneAnalysis results for data from the 
# two individuals included here in this vignette
cv233 <- system.file( "extdata/CV233_batchAnalysis.rds", package = "BrepPhylo")
cv325 <- system.file( "extdata/CV325_batchAnalysis.rds", package = "BrepPhylo")

# the getSummaryFromBatch function flattens the output to two
# elements: a table containing  distance-from-germline (demonstrated 
# above) and another table of all CSR events observed in lineages
cv233 <- getSummaryFromBatch( cv233 )
cv233 <- cv233$csr_events
cv325 <- getSummaryFromBatch( cv325 )
cv325 <- cv325$csr_events
# just add an extra column with PatientID
cv233$PatientID <- "CV233"
cv325$PatientID <- "CV325"
# we can combine the two tables
csr_details <- rbind( cv233, cv325 )
csr_summary <- summariseCSR( csr_details, 
                             dist_column = "distFromGermline", 
                             cloneID_column = "CloneID",
                             summarise_variables = "PatientID")
}

}
