% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/germlineDistances.R
\name{summariseGermlineDistance}
\alias{summariseGermlineDistance}
\title{Function to summarise distance-from-germline by clones}
\usage{
summariseGermlineDistance(
  distFromGermline,
  dist_column = "dist",
  cloneID_column = "CloneID",
  summarise_variables
)
}
\arguments{
\item{distFromGermline}{data.frame, output of \code{getGermlineDistance} and annotated with extra columns detailing metadata of samples.}

\item{dist_column}{Character, column name in \code{distFromGermline} which holds the distance-from-germline metrics.}

\item{cloneID_column}{Character, column name in \code{distFromGermline} which holds the clone ID to which the given sequence belongs.}

\item{summarise_variables}{Character vector, names of columns in \code{distFromGermline} used to partition the sequences into subsets for which summary is sought}
}
\value{
A data.frame with each row corresponding to one clone, containing, in addition to metadata columns given in \code{summarise_variables}, the following columns:
\describe{
  \item{dist_median}{Numeric, between 0 to 1, the percentile of the actual median distance-from-germline of the given clone.}
  \item{clone_order}{Numeric, between 0 to 1, the percentile of the given clone in the distribution where all clones are ordered by their median distance.}
}
}
\description{
The function first summarises for each clone (optionally with other metadata columns present in the input \code{distFromGermline} data frame) the median distance-from-germline of all sequences in the clone.
It then used this median distance to order the clones (from lowest - i.e. fewest mutations - to highest) and treat this as an 'expected' order of the clone in terms of mutational level, and calculate for each clone its 'actual' percentile of its median distance over all other clones.
The discrepancy between the 'actual' and 'expected' percentiles returned by this function can subsequently be used to plot curves showing overall mutational level of a repertoire (see Example), or to seek for a quantification using the Germline Likeness metric (see \code{?getGermlineLikeness}).
}
