% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Vgermline}
\alias{Vgermline}
\title{Collection of germline V DNA sequences}
\format{
a list of 6 entries named by the organism names listed in the description above (notice the underscore between binomial names). Each entry is a named list of character strings holding the sequences. V gene (e.g. IGHV1-1) are names of this list.
}
\usage{
Vgermline
}
\description{
ungapped V-gene DNA sequences taken November 2020 from the IMGT webserver, for the following species: Bos_taurus, Gallus_gallus, Homo_sapiens, Mus_musculus, Oryctolagus_cuniculus, Sus_scrofa.
}
\keyword{datasets}
